+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## Version 0.2.1-beta - 2022-10-22
+
+### Fixed
+
+- Fixed a number of HTML tools not echoing correctly.
+
+---
+
 ## Version 0.2.0-beta - 2022-10-05
 
 ### Added